Implement multiple-filesystem support for remote targets
This commit allows GDB to access executables and shared libraries
on remote targets where the inferior and the remote stub do not
share a common filesystem.  A new packet "vFile:setfs" is added
to the remote protocol and the three remote hostio functions with
filename arguments are modified to send "vFile:setfs" packets as
necessary.