Update how find_separate_debug_file handles CANON_DIR argument
find_separate_debug_file's CANON_DIR argument is used to create an
alternate location to search for debug files.  This commit moves
that logic out of the loop it's in and introduces a new variable
altdir to hold the generated alternate location.  A check is added
to inhibit the altdir search if alternate location is the same as
the regular location specified in DIR.

gdb/ChangeLog:

	* symfile.c (find_separate_debug_file): Move canon_dir
	logic out of loop.  Don't check the same location twice.