Pre-strip now-unnecessary trailing directory separators
Prior to the previous commit find_separate_debug_file required that
its DIR argument had a trailing directory separator.  This is no
longer necessary, and makes it difficult to check whether dir and
canon_dir are the same as canon_dir usually does not have a trailing
separator.  This commit updates find_separate_debug_file's caller to
not supply DIR with a trailing separator.  The next commit in the
series relies on this to avoid trying the same location twice.

gdb/ChangeLog:

	* symfile.c (find_separate_debug_file): Update comment.
	(terminate_after_last_dir_separator): Replaced with...
	(terminate_at_last_dir_separator): New function.
	(find_separate_debug_file_by_debuglink): Use the above.