Also look for debug files in gdb_sysroot
This commit makes GDB look for separate debug files in gdb_sysroot if
not found in the currently searched locations.  This allows for easier
analysis of core files from foreign machines using the sysroot option.
The user creates or mounts a directory containing the necessary
binaries and debuginfo and then sets GDB's sysroot to that directory
before starting their debug session.

gdb/ChangeLog:

	* symfile.c (find_separate_debug_file): In debugdir loop,
	also try alternate directory prefixed with gdb_sysroot.