Add "target:" filename handling to find_separate_debug_file
This commit updates find_separate_debug_file to handle filenames
prefixed with "target:".  The same-directory and DEBUG_SUBDIRECTORY
locations are checked with the prefix if supplied.  The debugdir
location is checked both with and without the prefix if one is
supplied.  This makes GDB able to fetch separate debug files from
remote targets and from inferiors in containers.

gdb/ChangeLog:

	* utils.h (build_filename): New argument "prefix".
	* utils.c (build_filename): Likewise.
	* symfile.c (find_separate_debug_file): Separate
	TARGET_SYSROOT_PREFIX from dir.  In debugdir loop, also try
	location prepended with dir's prefix if one was supplied.