Update exec_file_find callers
This commit updates the two callers of exec_file_find to make the call
regardless of what gdb_sysroot is set to.

Also, one of the callers used exec_file_find for absolute filenames
only, but the other used it in all cases.  The latter is probably
wrong, and has been updated to only use exec_file_find for absolute
filenames.

gdb/ChangeLog:

	* exec.c (exec_file_locate_attach): Use exec_file_find
	regardless of what gdb_sysroot is set to.
	* infrun.c (follow_exec): Likewise, but only if the new
	filename is absolute.